Work on duplicated code
