+++ app/src/main/java/com/github/meeplemeet/ui/ShopFormComponents.kt
@@ -1,0 +1,475 @@
+// This file contains shared components, test tags, and utilities used by both
+// CreateShopScreen and ShopDetailsEditScreen
+// Github copilot was used for this file
+
+package com.github.meeplemeet.ui
+
+import android.util.Patterns
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ExpandMore
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+import com.github.meeplemeet.model.sessions.Game
+import com.github.meeplemeet.model.shared.Location
+import com.github.meeplemeet.model.shops.OpeningHours
+import com.github.meeplemeet.model.shops.TimeSlot
+import com.github.meeplemeet.ui.components.DayRow
+import com.github.meeplemeet.ui.components.GameStockDialog
+import com.github.meeplemeet.ui.components.LabeledField
+import com.github.meeplemeet.ui.components.LocationSearchField
+import com.github.meeplemeet.ui.components.OpeningHoursDialog
+import java.text.DateFormatSymbols
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
+import java.util.Locale
+import kotlin.random.Random
+
+/* ================================================================================================
+ * Shared Test Tags
+ * ================================================================================================ */
+object ShopFormTestTags {
+  const val SECTION_HEADER_SUFFIX = "_header"
+  const val SECTION_TITLE_SUFFIX = "_title"
+  const val SECTION_TOGGLE_SUFFIX = "_toggle"
+  const val SECTION_DIVIDER_SUFFIX = "_divider"
+  const val SECTION_CONTENT_SUFFIX = "_content"
+
+  const val FIELD_SHOP = "field_shop_name"
+  const val FIELD_EMAIL = "field_email"
+  const val FIELD_ADDRESS = "field_address"
+  const val FIELD_PHONE = "field_phone"
+  const val FIELD_LINK = "field_link"
+
+  const val AVAILABILITY_LIST = "availability_list"
+  const val AVAILABILITY_DIVIDER_PREFIX = "availability_divider_"
+
+  const val OPENING_HOURS_DIALOG_WRAPPER = "opening_hours_dialog_wrapper"
+  const val GAME_STOCK_DIALOG_WRAPPER = "game_stock_dialog_wrapper"
+}
+
+/* ================================================================================================
+ * Shared UI Defaults
+ * ================================================================================================ */
+object ShopFormUi {
+  object Dimensions {
+    val contentHPadding = 16.dp
+    val contentVPadding = 8.dp
+    val sectionSpace = 12.dp
+    val bottomSpacer = 100.dp
+    val betweenControls = 6.dp
+  }
+
+  object Strings {
+    const val LabelShop = "Shop"
+    const val PlaceholderShop = "Shop name"
+
+    const val LabelEmail = "Email"
+    const val PlaceholderEmail = "Email"
+
+    const val LabelPhone = "Contact info"
+    const val PlaceholderPhone = "Phone number"
+
+    const val LabelLink = "Link"
+    const val PlaceholderLink = "Website/Instagram link"
+
+    const val PlaceholderLocation = "Search locations…"
+
+    const val Collapse = "Collapse"
+    const val Expand = "Expand"
+
+    const val ClosedMsg = "Closed"
+    const val Open24Msg = "Open 24 hours"
+  }
+
+  val dayNames: List<String> by lazy {
+    val weekdays = DateFormatSymbols().weekdays
+    (0..6).map { idx -> weekdays.getOrNull(idx + 1) ?: "Day ${idx + 1}" }
+  }
+}
+
+/* ================================================================================================
+ * Time utilities
+ * ================================================================================================ */
+object TimeUi {
+  const val OPEN24_START = "00:00"
+  const val OPEN24_END = "23:59"
+  val fmt12: DateTimeFormatter = DateTimeFormatter.ofPattern("h:mm a", Locale.getDefault())
+}
+
+/**
+ * Formats a LocalTime object into "HH:mm" string format.
+ *
+ * @return A string representation of the time in "HH:mm" format.
+ * @receiver The LocalTime object to format.
+ */
+fun LocalTime.hhmm(): String = "%02d:%02d".format(hour, minute)
+
+/**
+ * Tries to parse a time string into a LocalTime object.
+ *
+ * Supports both 12-hour (with AM/PM) and 24-hour formats.
+ *
+ * @return A LocalTime object if parsing is successful, null otherwise.
+ * @receiver The time string to parse.
+ */
+fun String.tryParseTime(): LocalTime? =
+    runCatching {
+          val lower = lowercase(Locale.getDefault())
+          if (lower.contains("am") || lower.contains("pm")) {
+            LocalTime.parse(
+                replace("am", " AM", ignoreCase = true)
+                    .replace("pm", " PM", ignoreCase = true)
+                    .replace(Regex("\\s+"), " ")
+                    .trim()
+                    .uppercase(Locale.getDefault()),
+                TimeUi.fmt12)
+          } else {
+            val (h, mRest) = split(":")
+            LocalTime.of(h.toInt(), mRest.take(2).toInt())
+          }
+        }
+        .getOrNull()
+
+/**
+ * Converts a list of TimeSlot objects into a human-readable string format.
+ *
+ * @param hours List of TimeSlot objects representing opening hours.
+ * @return A string representation of the opening hours.
+ */
+fun humanize(hours: List<TimeSlot>): String =
+    when {
+      hours.isEmpty() -> ShopFormUi.Strings.ClosedMsg
+      hours.size == 1 &&
+          hours[0].open == TimeUi.OPEN24_START &&
+          hours[0].close == TimeUi.OPEN24_END -> ShopFormUi.Strings.Open24Msg
+      else -> {
+        // Sort by opening time
+        val sorted = hours.sortedBy { ts -> ts.open?.tryParseTime() ?: LocalTime.MAX }
+        sorted.joinToString("\n") { slot ->
+          val s = slot.open?.let { it.tryParseTime()?.format(TimeUi.fmt12) ?: it } ?: "-"
+          val e = slot.close?.let { it.tryParseTime()?.format(TimeUi.fmt12) ?: it } ?: "-"
+          "$s - $e"
+        }
+      }
+    }
+
+/* ================================================================================================
+ * Helpers
+ * ================================================================================================ */
+
+/** Returns an empty list of opening hours for each day of the week. */
+fun emptyWeek(): List<OpeningHours> =
+    List(7) { day -> OpeningHours(day = day, hours = emptyList()) }
+
+/**
+ * Mocks location suggestions based on a query string. Kept local and private for simplicity until a
+ * real provider replaces it.
+ */
+private fun mockLocationSuggestionsFrom(query: String, max: Int = 5): List<Location> {
+  if (query.isBlank()) return emptyList()
+  val rng = Random(query.hashCode())
+  return List(max) { i ->
+    val lat = rng.nextDouble(-90.0, 90.0)
+    val lon = rng.nextDouble(-180.0, 180.0)
+    Location(latitude = lat, longitude = lon, name = "$query #${i + 1}")
+  }
+}
+
+/**
+ * Validates if the provided email string is in a valid email format.
+ *
+ * @param email The email string to validate.
+ * @return True if the email is valid, false otherwise.
+ */
+fun isValidEmail(email: String): Boolean = Patterns.EMAIL_ADDRESS.matcher(email).matches()
+
+/* ================================================================================================
+ * Shared Composable Components
+ * ================================================================================================ */
+
+/**
+ * Composable function representing the required information section.
+ *
+ * @param shopName The current value of the shop name field.
+ * @param onShopName Callback function to update the shop name.
+ * @param email The current value of the email field.
+ * @param onEmail Callback function to update the email.
+ * @param phone The current value of the phone field.
+ * @param onPhone Callback function to update the phone.
+ * @param link The current value of the link field.
+ * @param onLink Callback function to update the link.
+ * @param addressText The current value of the address text field.
+ * @param onAddressText Callback function to update the address text.
+ * @param onPickLocation Callback function to handle location selection.
+ */
+@Composable
+fun RequiredInfoSection(
+    shopName: String,
+    onShopName: (String) -> Unit,
+    email: String,
+    onEmail: (String) -> Unit,
+    phone: String,
+    onPhone: (String) -> Unit,
+    link: String,
+    onLink: (String) -> Unit,
+    addressText: String,
+    onAddressText: (String) -> Unit,
+    onPickLocation: (Location) -> Unit
+) {
+  Box(Modifier.testTag(ShopFormTestTags.FIELD_SHOP)) {
+    LabeledField(
+        label = ShopFormUi.Strings.LabelShop,
+        placeholder = ShopFormUi.Strings.PlaceholderShop,
+        value = shopName,
+        onValueChange = onShopName)
+  }
+  Box(Modifier.testTag(ShopFormTestTags.FIELD_EMAIL)) {
+    LabeledField(
+        label = ShopFormUi.Strings.LabelEmail,
+        placeholder = ShopFormUi.Strings.PlaceholderEmail,
+        value = email,
+        onValueChange = onEmail,
+        keyboardType = KeyboardType.Email)
+  }
+  val showEmailError = email.isNotEmpty() && !isValidEmail(email)
+  if (showEmailError) {
+    Text(
+        "Enter a valid email address.",
+        color = MaterialTheme.colorScheme.error,
+        style = MaterialTheme.typography.bodySmall)
+  }
+  Box(Modifier.testTag(ShopFormTestTags.FIELD_PHONE)) {
+    LabeledField(
+        label = ShopFormUi.Strings.LabelPhone,
+        placeholder = ShopFormUi.Strings.PlaceholderPhone,
+        value = phone,
+        onValueChange = onPhone,
+        keyboardType = KeyboardType.Phone)
+  }
+
+  Box(Modifier.testTag(ShopFormTestTags.FIELD_LINK)) {
+    LabeledField(
+        label = ShopFormUi.Strings.LabelLink,
+        placeholder = ShopFormUi.Strings.PlaceholderLink,
+        value = link,
+        onValueChange = onLink,
+        keyboardType = KeyboardType.Uri)
+  }
+
+  Box(Modifier.testTag(ShopFormTestTags.FIELD_ADDRESS)) {
+    val locationResults = remember(addressText) { mockLocationSuggestionsFrom(addressText) }
+    LocationSearchField(
+        query = addressText,
+        onQueryChange = onAddressText,
+        results = locationResults,
+        onPick = onPickLocation,
+        isLoading = false,
+        placeholder = ShopFormUi.Strings.PlaceholderLocation,
+        modifier = Modifier.fillMaxWidth())
+  }
+}
+
+/**
+ * Composable function representing the availability section.
+ *
+ * @param week List of opening hours for each day of the week.
+ * @param onEdit Callback function to handle editing of opening hours for a specific day.
+ */
+@Composable
+fun AvailabilitySection(week: List<OpeningHours>, onEdit: (Int) -> Unit) {
+  Column(Modifier.testTag(ShopFormTestTags.AVAILABILITY_LIST)) {
+    week.forEach { oh ->
+      val day = oh.day
+      DayRow(
+          dayName = ShopFormUi.dayNames[day], value = humanize(oh.hours), onEdit = { onEdit(day) })
+      HorizontalDivider(
+          modifier = Modifier.testTag(ShopFormTestTags.AVAILABILITY_DIVIDER_PREFIX + day))
+    }
+  }
+  Spacer(Modifier.height(4.dp))
+}
+
+/**
+ * Composable function representing a collapsible section with a title, optional header, and
+ * content.
+ *
+ * @param title The title of the section.
+ * @param initiallyExpanded Boolean indicating whether the section is initially expanded.
+ * @param header Optional composable function for the header content.
+ * @param content Composable function for the main content of the section.
+ * @param testTag Optional test tag for the section.
+ */
+@Composable
+fun CollapsibleSection(
+    title: String,
+    initiallyExpanded: Boolean = true,
+    header: (@Composable RowScope.() -> Unit)? = null,
+    content: @Composable ColumnScope.() -> Unit,
+    testTag: String? = null
+) {
+  var expanded by rememberSaveable { mutableStateOf(initiallyExpanded) }
+  val arrowRotation by animateFloatAsState(if (expanded) 180f else 0f, label = "arrow")
+
+  Column(Modifier.fillMaxWidth()) {
+    Row(
+        modifier =
+            Modifier.fillMaxWidth().padding(top = 8.dp).let { m ->
+              if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_HEADER_SUFFIX)
+              else m
+            },
+        verticalAlignment = Alignment.CenterVertically) {
+          Text(
+              text = title,
+              style = MaterialTheme.typography.titleMedium,
+              modifier =
+                  Modifier.weight(1f).let { m ->
+                    if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_TITLE_SUFFIX)
+                    else m
+                  })
+          header?.invoke(this)
+          IconButton(
+              onClick = { expanded = !expanded },
+              modifier =
+                  Modifier.let { m ->
+                    if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_TOGGLE_SUFFIX)
+                    else m
+                  }) {
+                Icon(
+                    Icons.Filled.ExpandMore,
+                    contentDescription =
+                        if (expanded) ShopFormUi.Strings.Collapse else ShopFormUi.Strings.Expand,
+                    modifier = Modifier.rotate(arrowRotation))
+              }
+        }
+    HorizontalDivider(
+        thickness = 1.dp,
+        color = MaterialTheme.colorScheme.outlineVariant,
+        modifier =
+            Modifier.padding(bottom = 12.dp).let { m ->
+              if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_DIVIDER_SUFFIX)
+              else m
+            })
+
+    AnimatedVisibility(visible = expanded) {
+      Column(
+          Modifier.padding(top = 0.dp).let { m ->
+            if (testTag != null) m.testTag(testTag + ShopFormTestTags.SECTION_CONTENT_SUFFIX) else m
+          },
+          content = content)
+    }
+  }
+}
+
+/**
+ * Composable function representing the opening hours editor dialog.
+ *
+ * @param show Boolean indicating whether to show the dialog.
+ * @param day The day of the week being edited.
+ * @param week List of opening hours for each day of the week.
+ * @param onWeekChange Callback function to update the opening hours for the week.
+ * @param onDismiss Callback function to dismiss the dialog.
+ */
+@Composable
+fun OpeningHoursEditor(
+    show: Boolean,
+    day: Int?,
+    week: List<OpeningHours>,
+    onWeekChange: (List<OpeningHours>) -> Unit,
+    onDismiss: () -> Unit
+) {
+  if (!show || day == null) return
+  Box(Modifier.testTag(ShopFormTestTags.OPENING_HOURS_DIALOG_WRAPPER)) {
+    OpeningHoursDialog(
+        initialSelectedDays = setOf(day),
+        current = week[day],
+        onDismiss = onDismiss,
+        onSave = { selectedDays, closed, open24, intervals ->
+          val encoded: List<TimeSlot> =
+              when {
+                closed -> emptyList()
+                open24 -> listOf(TimeSlot(TimeUi.OPEN24_START, TimeUi.OPEN24_END))
+                else -> intervals.map { TimeSlot(it.first.hhmm(), it.second.hhmm()) }
+              }
+          val copy = week.toMutableList()
+          selectedDays.forEach { d -> copy[d] = OpeningHours(day = d, hours = encoded) }
+          onWeekChange(copy)
+          onDismiss()
+        })
+  }
+}
+
+/**
+ * Composable function representing the game stock picker dialog.
+ *
+ * @param show Boolean indicating whether to show the dialog.
+ * @param stock List of pairs containing games and their quantities in stock.
+ * @param onStockChange Callback function to update the stock list.
+ * @param gameQuery The current query string for searching games.
+ * @param gameSuggestions List of game suggestions based on the current query.
+ * @param isSearching Boolean indicating if a search operation is in progress.
+ * @param picked The currently picked game.
+ * @param onPickedChange Callback function to update the picked game.
+ * @param qty The quantity of the picked game.
+ * @param onQtyChange Callback function to update the quantity of the picked game.
+ * @param onSetGameQuery Callback function to update the game search query.
+ * @param onSetGame Callback function to set the selected game.
+ * @param onDismiss Callback function to dismiss the dialog.
+ */
+@Composable
+fun GameStockPicker(
+    show: Boolean,
+    stock: List<Pair<Game, Int>>,
+    onStockChange: (List<Pair<Game, Int>>) -> Unit,
+    gameQuery: String,
+    gameSuggestions: List<Game>,
+    isSearching: Boolean,
+    picked: Game?,
+    onPickedChange: (Game?) -> Unit,
+    qty: Int,
+    onQtyChange: (Int) -> Unit,
+    onSetGameQuery: (String) -> Unit,
+    onSetGame: (Game) -> Unit,
+    onDismiss: () -> Unit
+) {
+  if (!show) return
+
+  val existing = remember(stock) { stock.map { it.first.uid }.toSet() }
+  Box(Modifier.testTag(ShopFormTestTags.GAME_STOCK_DIALOG_WRAPPER)) {
+    GameStockDialog(
+        query = gameQuery,
+        onQueryChange = onSetGameQuery,
+        results = gameSuggestions,
+        isLoading = isSearching,
+        onPickGame = { g ->
+          onPickedChange(g)
+          onSetGame(g)
+        },
+        selectedGame = picked,
+        quantity = qty,
+        onQuantityChange = onQtyChange,
+        existingIds = existing,
+        onDismiss = {
+          onDismiss()
+          onQtyChange(1)
+          onPickedChange(null)
+          onSetGameQuery("")
+        },
+        onSave = {
+          picked?.let { g -> onStockChange((stock + (g to qty)).distinctBy { it.first.uid }) }
+          onQtyChange(1)
+          onPickedChange(null)
+          onSetGameQuery("")
+          onDismiss()
+        })
+  }
+}
